MISC: 코드 정리, Network 처리를 어디서 해야할 지 고민해야함 (Scene의 Update 순서와 상관없이 recv가 호출된다.)

+++ ConSoleStudy2023/NetworkMGR.h
@@ -21,7 +21,7 @@ public:
 	char* SERVERIP;
 	shared_ptr<Socket> tcpSocket;
 
-	int id;
+	unsigned long long id;
 	string name;
 	bool is_mage;
